Persuade add_unittest to really add tests.
In a standalone build the variable LLVM_MAIN_SRC_DIR is not set.
An override of function add_unittest sets this variable to the
base source folder, because the only use is to add the Google Test
headers. With this change it is possible to compile unittests
in a standaone build.