Fix a bug in LLtool with alternatives.
For alternatives which can derive epsilon, the wrong condition was
generated. Instead of only using the follow set in case the first set
is empty, now the union of first and follow set is used.
To avoid long if statements (follow sets can be large), a check is
added if this alternative is the last one. In this case, the condition
is dropped and only an else statement is generated.
Dropping the first check on a sequence is updated to take into
account that the sequence could derive epsilon, in which case
the check cannot be dropped. This may need more refinement.